Added more to the UI.  There is a stupid bug in adding actors in the MovieAddEditDialog.  Mess around with it and you will quickly see why.  I have not figured it out yet but I will.
The image in the ItemInfoPanel is a placeholder.  It doesn't look to bad though.

Added new, better icons.

Search text field is better now.  It will not remove text in the text field when focus is lost.

MovieAddEditDialog has been implemented.  Still needs work.

Minor adjustments to component alignment and placement.

git-svn-id: http://svn.assembla.com/svn/mangoproject/trunk@63 e0e46c49-be69-4f5a-ad62-21024a331aea