Javadoc'd ui package, except Mango because it is a big baby if you do anything to it.  It is like an innocent home school girl that stays in her room and reads books on poetry.
Updated MovieAddEditDialog to save role and character, but we can't update the role and character, so messed up but whatever.

Added a button to the bottom panel so we can set the owner.  I tried to make it a menu item but of course mango is a girl and started crying.

git-svn-id: http://svn.assembla.com/svn/mangoproject/trunk@303 e0e46c49-be69-4f5a-ad62-21024a331aea